task:  move global directory content into PokerGame container component's feature folder to co-locate the files
