+++ src/components/containers/PokerGame/features/deck-dealers/dealers/RandomDealer.ts
@@ -1,4 +1,4 @@
-import type { CardDeck } from "../../types";
+import type { CardDeck } from "../../../../../../global/types";
 import type { DeckDealer } from "../types";
 
 export default class RandomDealer implements DeckDealer {
