+++ src/components/containers/PokerGame/features/deck-dealers/types.ts
@@ -1,4 +1,4 @@
-import type { Hand } from "../types";
+import type { Hand } from "../../../../../global/types";
 
 export interface DeckDealer {
   deal: () => Hand;
