+++ src/components/containers/PokerGame/features/hand-evaluators/evaluators/FourOfAKindEvaluator.ts
@@ -1,17 +1,16 @@
 import type { HandEvaluator } from "../types";
-import type { Hand } from "../../types";
+import type { Hand } from "../../../../../../global/types";
 
 interface Frequencies {
   [key: string]: any;
 }
-
-export default class PairEvaluator implements HandEvaluator {
+export default class FourOfAKindEvaluator implements HandEvaluator {
   constructor(private hand: Hand) {}
 
   public evaluate() {
     const { hand } = this;
     const frequencies: Frequencies = {};
-    let isPair = false;
+    let isFourOfKind = false;
 
     for (const card of hand) {
       const cardProperty = card.number;
@@ -21,11 +20,11 @@ export default class PairEvaluator implements HandEvaluator {
     for (const [cardNumber, cardNumberFrequency] of Object.entries(
       frequencies
     )) {
-      if (cardNumberFrequency === 2) {
-        isPair = true;
+      if (cardNumberFrequency === 4) {
+        isFourOfKind = true;
       }
     }
 
-    return isPair;
+    return isFourOfKind;
   }
 }
