+++ src/components/containers/PokerGame/features/hand-evaluators/evaluators/StraighFlushEvaluator.ts
@@ -1,5 +1,5 @@
 import type { HandEvaluator } from "../types";
-import type { Hand } from "../../types";
+import type { Hand } from "../../../../../../global/types";
 
 export default class StraightFlushEvaluator implements HandEvaluator {
   constructor(private hand: Hand) {}
