+++ src/components/containers/PokerGame/features/shufflers/algorithms/RandomShuffler.ts
@@ -1,4 +1,4 @@
-import type { CardDeck } from "../../types";
+import type { CardDeck } from "../../../../../../global/types";
 import type { DeckShuffler } from "../types";
 
 export default class RandomShuffler implements DeckShuffler {
